consolidate the four GTM plugins into one `marketing` plugin
build-in-public, social-posting, x-posting, and outreach were four
overlapping "post about ourselves / reach out" plugins — cruft in the
inventory that reads as four products to a stranger. Merged into a single
`marketing` plugin.

The consolidation is in the PACKAGING, not the surface: each area keeps its
own routes/db/helpers in a subdirectory (marketing/{bip,social,x,outreach}),
and the plugin mounts them at their original sub-prefixes from routePrefix
"/". So every external path (/bip, /social, /x, /outreach), all 25 MCP tool
names and their endpoints, and all six table names are UNCHANGED — the live
X tools (x_publish, etc.) the squad uses keep working, no data migration.

- One plugin.json (one enable toggle), gatedActions union of all four
  (bip_post_publish, outreach_send, social_publish, x_publish), X config
  schema preserved.
- schema.sql + mcp-tools.json concatenated from the four areas; redundant
  area-level manifests removed.
- social area's core-db import fixed for the extra nesting level.

Plugin count 17 → 13 (with the SaaS removal). Boots clean, 147 tests pass,
all four sub-routes resolve (401 not 404), marketing loads with 25 MCP tools.
