+++ server/plugins/marketing/outreach/lib/researcher.js
