--- server/plugins/social-posting/mcp-tools.json
@@ -1,101 +1,0 @@
-[
-  {
-    "name": "mycelium_social_generate_caption",
-    "description": "Generate a character-voiced social media caption prompt for a gameplay highlight clip. Returns system prompt + user prompt to send to Claude API.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "event_type": { "type": "string", "description": "Highlight event type (perfect_roll, mythic_drop, player_death, etc.)" },
-        "metadata": { "type": "object", "description": "Event metadata (quality_score, item_name, damage, floor, etc.)" },
-        "platform": { "type": "string", "enum": ["tiktok", "instagram", "twitter", "youtube_shorts"], "description": "Target platform" }
-      },
-      "required": ["event_type", "platform"]
-    },
-    "endpoint": { "method": "POST", "path": "/social/captions/generate", "bodyMap": { "event_type": "event_type", "metadata": "metadata", "platform": "platform" } }
-  },
-  {
-    "name": "mycelium_social_create_post",
-    "description": "Create a social media post (draft or scheduled). Link to a video pipeline clip for full tracking.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "project_id": { "type": "string" },
-        "platform": { "type": "string", "enum": ["tiktok", "instagram", "twitter", "youtube_shorts"] },
-        "caption": { "type": "string", "description": "Post caption text" },
-        "media_url": { "type": "string", "description": "URL of the video/image to post" },
-        "clip_id": { "type": "string", "description": "Video pipeline clip ID (for tracking)" },
-        "video_session_id": { "type": "integer", "description": "Video pipeline session ID" },
-        "event_type": { "type": "string" },
-        "tier": { "type": "string", "enum": ["S", "A", "B", "C"] },
-        "status": { "type": "string", "enum": ["draft", "scheduled"], "description": "Initial status (default: draft)" },
-        "scheduled_at": { "type": "string", "description": "ISO timestamp for scheduled posting" }
-      },
-      "required": ["project_id", "platform"]
-    },
-    "endpoint": { "method": "POST", "path": "/social/posts", "bodyMap": { "project_id": "project_id", "platform": "platform", "caption": "caption", "media_url": "media_url", "clip_id": "clip_id", "video_session_id": "video_session_id", "event_type": "event_type", "tier": "tier", "status": "status", "scheduled_at": "scheduled_at" } }
-  },
-  {
-    "name": "mycelium_social_list_posts",
-    "description": "List social media posts with optional filters.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "project_id": { "type": "string" },
-        "platform": { "type": "string" },
-        "status": { "type": "string", "enum": ["draft", "scheduled", "publishing", "posted", "failed"] },
-        "video_session_id": { "type": "integer" }
-      }
-    },
-    "endpoint": { "method": "GET", "path": "/social/posts", "queryMap": { "project_id": "project_id", "platform": "platform", "status": "status", "video_session_id": "video_session_id" } }
-  },
-  {
-    "name": "mycelium_social_schedule_post",
-    "description": "Schedule a draft post for publishing. Auto-selects next available time window if no scheduled_at provided.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "post_id": { "type": "integer", "description": "Post ID to schedule" },
-        "scheduled_at": { "type": "string", "description": "ISO timestamp (optional — auto-schedules if omitted)" }
-      },
-      "required": ["post_id"]
-    },
-    "endpoint": { "method": "POST", "path": "/social/posts/{post_id}/schedule", "bodyMap": { "scheduled_at": "scheduled_at" } }
-  },
-  {
-    "name": "mycelium_social_publish_post",
-    "description": "Publish a post to its platform via drone job. Requires caption and media_url. Uses Buffer API for TikTok/Twitter or Instagram Graph API. Gated action: social_publish.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "post_id": { "type": "integer", "description": "Post ID to publish" },
-        "priority": { "type": "integer" }
-      },
-      "required": ["post_id"]
-    },
-    "endpoint": { "method": "POST", "path": "/social/posts/{post_id}/publish", "bodyMap": { "priority": "priority" } }
-  },
-  {
-    "name": "mycelium_social_stats",
-    "description": "Get posting statistics by platform and status.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "project_id": { "type": "string", "description": "Project ID" }
-      }
-    },
-    "endpoint": { "method": "GET", "path": "/social/stats", "queryMap": { "project_id": "project_id" } }
-  },
-  {
-    "name": "mycelium_social_history",
-    "description": "Get recent post history for scheduling decisions.",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "project_id": { "type": "string" },
-        "platform": { "type": "string" },
-        "days": { "type": "integer", "description": "Number of days of history (default: 7)" }
-      }
-    },
-    "endpoint": { "method": "GET", "path": "/social/history", "queryMap": { "project_id": "project_id", "platform": "platform", "days": "days" } }
-  }
-]
